move stuff to resolvers.py and schema.py
