+++ python/whylogs/core/resolvers.py
@@ -1,11 +1,16 @@
+import logging
 from abc import ABC, abstractmethod
-from typing import Dict, List, TypeVar
+from dataclasses import dataclass, field
+from typing import Any, Dict, List, Optional, TypeVar
 
 from typing_extensions import TypeAlias
 
 from whylogs.core.datatypes import DataType, Fractional, Integral, String
 from whylogs.core.metrics import StandardMetric
-from whylogs.core.metrics.metrics import Metric
+from whylogs.core.metrics.metrics import Metric, MetricConfig
+
+logger = logging.getLogger(__name__)
+
 
 M = TypeVar("M", bound=Metric)
 ColumnSchema: TypeAlias = "ColumnSchema"  # type: ignore
@@ -78,4 +83,81 @@ class HistogramCountingTrackingResolver(Resolver):
         result: Dict[str, Metric] = {}
         for m in metrics:
             result[m.name] = m.zero(column_schema.cfg)
+        return result
+
+
+@dataclass
+class MetricSpec:
+    """
+    Specify a Metric to instantiate.
+    """
+
+    metric: Any  # Should be a subclass of Metric, it should be the class, not an instance
+    config: Optional[MetricConfig] = None  # omit to use default MetricConfig
+
+
+@dataclass
+class ResolverSpec:
+    """
+    Specify the metrics to instantiate for matching columns. column_name
+    takes precedence over column_type. column_type should be a subclass
+    of DataType, i.e., AnyType, Frational, Integral, or String. Pass the
+    class, not an instance.
+    """
+
+    column_name: Optional[str] = None  # TODO: maybe make this a regex
+    column_type: Optional[Any] = None
+    metrics: List[MetricSpec] = field(default_factory=list)
+
+
+# whylabs expects COLUMN_METRICS to be present for every column.
+COLUMN_METRICS = [MetricSpec(StandardMetric.counts.value), MetricSpec(StandardMetric.types.value)]
+
+
+def _allowed_metric(config: MetricConfig, metric: Metric) -> bool:
+    """Return False for any metrics turned off in the config"""
+
+    namespace = metric.get_namespace()
+    if config.fi_disabled and namespace == "frequent_items":
+        return False
+    if (not config.track_unicode_ranges) and namespace == "unicode_range":
+        return False
+    return True
+
+
+class DeclarativeResolver(Resolver):
+    """
+    Implements the declarative resolution logic by interpreting a "program"
+    of ResolverSpecs
+    """
+
+    def __init__(self, resolvers: List[ResolverSpec], default_config: Optional[MetricConfig] = None) -> None:
+        # Validate resolvers -- must have name xor type, MetricSpec metrcis must <: Metric
+        for spec in resolvers:
+            if spec.column_name and spec.column_type:
+                logger.warning(
+                    f"DeclarativeSchema: column {spec.column_name} also specified type, name takes precedence"
+                )
+            if not (spec.column_name or spec.column_type):
+                raise ValueError("DeclarativeSchema: resolver specification must supply name or type")
+
+            if spec.column_type and not issubclass(spec.column_type, DataType):
+                raise ValueError("DeclarativeSchema: resolver specification column type must be a DataType")
+
+            for metric_spec in spec.metrics:
+                if not issubclass(metric_spec.metric, Metric):
+                    raise ValueError("DeclarativeSchema: must supply a Metric subclass to MetricSpec")
+
+        self._resolvers = resolvers
+
+    def resolve(self, name: str, why_type: DataType, column_schema: ColumnSchema) -> Dict[str, Metric]:
+        result: Dict[str, Metric] = {}
+        for resolver_spec in self._resolvers:
+            col_name, col_type = resolver_spec.column_name, resolver_spec.column_type
+            if (col_name and col_name == name) or (col_name is None and isinstance(why_type, col_type)):  # type: ignore
+                for spec in resolver_spec.metrics:
+                    config = spec.config or column_schema.cfg
+                    if _allowed_metric(config, spec.metric):
+                        result[spec.metric.get_namespace()] = spec.metric.zero(config)
+
         return result
